+++ postcss.config.js
@@ -5,7 +5,7 @@ module.exports = (ctx) => ({
       root: 'src/css'
     } : false,
     'postcss-cssnext': ctx.env !== 'minify' ? {
-      browsers: ['last 2 versions', 'iOS >= 8']
+      browsers: ['last 2 versions', 'IE >= 10', 'Safari >= 6', 'iOS >= 5', '> 1%', 'Android >= 3']
     } : false,
     'cssnano': ctx.env === 'minify' ? {
       preset: ['default', {