+++ rollup.config.js
@@ -7,39 +7,64 @@ let today = new Date()
 let month = (today.getMonth() + 1).toString().padStart(2, '0')
 let day = today.getDate().toString().padStart(2, '0')
 let year = today.getFullYear()
+let banner = `/*! ${pkg.title || pkg.name} - v${pkg.version} - ${year}-${month}-${day}
+  ${pkg.homepage}
+  * Copyright (c) 2013-${year} ${pkg.author.name}; Licensed ${pkg.license} */`
 
-let baseConfig = {
-  banner: `/*! ${pkg.title || pkg.name} - v${pkg.version} - ${year}-${month}-${day}
-    ${pkg.homepage}
-    * Copyright (c) 2013-${year} ${pkg.author.name}; Licensed ${pkg.license} */`,
+export default [{
+  banner,
   input: 'src/jquery.sidr.js',
   globals: {
     jquery: 'jQuery'
-  }
-}
-
-let unminifiedConfig = Object.assign({}, baseConfig)
-unminifiedConfig.output = {
-  format: 'iife',
-  file: './dist/jquery.sidr.js'
-}
-unminifiedConfig.plugins = [
-  babel(babelrc())
-]
-
-let minifiedConfig = Object.assign({}, baseConfig)
-minifiedConfig.output = {
-  format: 'iife',
-  file: './dist/jquery.sidr.min.js'
-}
-minifiedConfig.plugins = [
-  babel(babelrc()),
-  uglify({
-    output: {
-      comments: /^!/
-    }
-  })
-]
-minifiedConfig.plugins.push()
-
-export default [unminifiedConfig, minifiedConfig]
+  },
+  output: {
+    format: 'iife',
+    file: './dist/jquery.sidr.js'
+  },
+  plugins: [
+    babel(babelrc())
+  ]
+}, {
+  banner,
+  input: 'src/jquery.sidr.js',
+  globals: {
+    jquery: 'jQuery'
+  },
+  output: {
+    format: 'iife',
+    file: './dist/jquery.sidr.min.js'
+  },
+  plugins: [
+    babel(babelrc()),
+    uglify({
+      output: {
+        comments: /^!/
+      }
+    })
+  ]
+}, {
+  banner,
+  input: 'src/sidr.js',
+  output: {
+    format: 'iife',
+    file: './dist/sidr.js'
+  },
+  plugins: [
+    babel(babelrc())
+  ]
+}, {
+  banner,
+  input: 'src/sidr.js',
+  output: {
+    format: 'iife',
+    file: './dist/sidr.min.js'
+  },
+  plugins: [
+    babel(babelrc()),
+    uglify({
+      output: {
+        comments: /^!/
+      }
+    })
+  ]
+}]