+++ src/js/menu.store.js
@@ -1,0 +1,11 @@
+let store = {}
+
+export default {
+  add (key, value) {
+    store[key] = value
+  },
+
+  get (key) {
+    return store[key]
+  }
+}